# lib/api-zod/src/generated/types/item.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Item {
  id: number;
  name: string;
  quantity: number;
  unit: string;
  updatedAt: Date;
  createdAt: Date;
}