# lib/api-zod/src/generated/types/updateItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateItemBody {
  /** @minimum 0 */
  quantity?: number;
  /**
   * @minLength 1
   * @maxLength 100
   */
  name?: string;
  /**
   * @minLength 1
   * @maxLength 30
   */
  unit?: string;
}